Carry out major code refactoring
So that the code is a lot lighter and the API matches the single
instantiation user requirements a bit more. This is achieved mainly by
restructuring the getter and setter methods. The user now can use and
reuse a single hash with key/values to set up the canvasloader instance,
Retrieving the key values the user can use `get('key')`.

- Introduce arguments hash for default key/value pairs
- Remove obsolete @method lines from code
- Replace `getValue` type getters with `get(key)` getters
- Remove setters completely and implemented initialiser hash object
- Introduce private methods using `_` prefix